test(config): split board tests.rs under 1000-LOC gate
board/tests.rs was 1025 LOC, tripping the repo's "reject .rs files
over 1000 LOC" CI gate. Move the data-driven enriched-JSON tests into
a sibling tests_enriched_json.rs (matching the existing tests_*.rs
convention) so each module stays under the limit. Pure code move; all
86 board tests still pass.
