+++ crates/fbuild-config/src/board/tests_enriched_json.rs
@@ -1,0 +1,671 @@
+//! Data-driven enriched JSON board-definition tests.
+//!
+//! Split out of `tests.rs` to keep each test module under the repo's
+//! 1000-LOC-per-file limit.
+
+use std::collections::HashMap;
+use std::io::Write;
+
+use tempfile::NamedTempFile;
+
+use super::db::get_board_db;
+use super::{BoardConfig, Esp32QemuPsramConfig};
+
+fn write_boards_txt(content: &str) -> NamedTempFile {
+    let mut f = NamedTempFile::new().unwrap();
+    f.write_all(content.as_bytes()).unwrap();
+    f.flush().unwrap();
+    f
+}
+
+#[test]
+fn test_mega_upload_protocol_wiring() {
+    // Bug fix: mega needs protocol=wiring, not arduino
+    let config = BoardConfig::from_board_id("mega", &HashMap::new()).unwrap();
+    assert_eq!(config.upload_protocol, Some("wiring".to_string()));
+    assert_eq!(config.core, "arduino");
+    assert_eq!(config.variant, "mega");
+}
+
+#[test]
+fn test_nano_upload_speed_57600() {
+    // Bug fix: nano uses 57600, not 115200
+    let config = BoardConfig::from_board_id("nano", &HashMap::new()).unwrap();
+    assert_eq!(config.upload_speed, Some("57600".to_string()));
+    assert_eq!(config.upload_protocol, Some("arduino".to_string()));
+    assert_eq!(config.variant, "eightanaloginputs");
+}
+
+#[test]
+fn test_teensy36_core_teensy3() {
+    // Bug fix: teensy30-36 use core=teensy3, not teensy4
+    let config = BoardConfig::from_board_id("teensy36", &HashMap::new()).unwrap();
+    assert_eq!(config.core, "teensy3");
+    assert_eq!(config.upload_protocol, Some("teensy-gui".to_string()));
+}
+
+#[test]
+fn test_teensy41_core_teensy4() {
+    let config = BoardConfig::from_board_id("teensy41", &HashMap::new()).unwrap();
+    assert_eq!(config.core, "teensy4");
+    assert_eq!(config.upload_protocol, Some("teensy-gui".to_string()));
+}
+
+/// Regression for FastLED/fbuild#300: every ARM-based Teensy board must
+/// declare its Teensyduino-bundled CMSIS-DSP math library via
+/// `build.cmsis_dsp_lib` so the linker can auto-link it (mirroring
+/// PlatformIO+Teensyduino's SCons behaviour). Without this, FastLED's
+/// `Ports/PJRCSpectrumAnalyzer` example and any sketch using `Audio.h`
+/// FFT classes fails with `undefined reference to arm_cfft_radix4_q15`.
+#[test]
+fn test_teensy_boards_declare_cmsis_dsp_lib() {
+    let cases: &[(&str, &str)] = &[
+        ("teensy30", "arm_cortexM4l_math"),
+        ("teensy31", "arm_cortexM4l_math"),
+        ("teensy35", "arm_cortexM4lf_math"),
+        ("teensy36", "arm_cortexM4lf_math"),
+        ("teensy40", "arm_cortexM7lfsp_math"),
+        ("teensy41", "arm_cortexM7lfsp_math"),
+        ("teensymm", "arm_cortexM7lfsp_math"),
+        ("teensylc", "arm_cortexM0l_math"),
+    ];
+    for (board_id, expected) in cases {
+        let config = BoardConfig::from_board_id(board_id, &HashMap::new()).unwrap();
+        assert_eq!(
+            config.cmsis_dsp_lib.as_deref(),
+            Some(*expected),
+            "{} should declare cmsis_dsp_lib={} (see fbuild#300)",
+            board_id,
+            expected
+        );
+    }
+}
+
+/// AVR-based Teensy 2.0 / Teensy++ 2.0 ship with the avr-libc math lib but
+/// no CMSIS-DSP — they're not Cortex-M parts. Their JSONs omit the field.
+#[test]
+fn test_teensy_avr_boards_no_cmsis_dsp_lib() {
+    for board_id in ["teensy2", "teensy2pp"] {
+        let config = BoardConfig::from_board_id(board_id, &HashMap::new()).unwrap();
+        assert_eq!(
+            config.cmsis_dsp_lib, None,
+            "{} is AVR-based and must not declare a CMSIS-DSP lib",
+            board_id
+        );
+    }
+}
+
+/// Non-Teensy boards must not have a CMSIS-DSP lib populated.
+#[test]
+fn test_non_teensy_boards_have_no_cmsis_dsp_lib() {
+    for board_id in ["uno", "mega", "esp32dev", "rpipico"] {
+        let config = BoardConfig::from_board_id(board_id, &HashMap::new()).unwrap();
+        assert_eq!(
+            config.cmsis_dsp_lib, None,
+            "{} should not declare a cmsis_dsp_lib",
+            board_id
+        );
+    }
+}
+
+/// Overrides (e.g. from `[env]` `board_build.cmsis_dsp_lib = ...`) should
+/// win over the bundled JSON value.
+#[test]
+fn test_cmsis_dsp_lib_override_wins() {
+    let mut overrides = HashMap::new();
+    overrides.insert(
+        "cmsis_dsp_lib".to_string(),
+        "arm_cortexM4lf_math_custom".to_string(),
+    );
+    let config = BoardConfig::from_board_id("teensy36", &overrides).unwrap();
+    assert_eq!(
+        config.cmsis_dsp_lib,
+        Some("arm_cortexM4lf_math_custom".to_string())
+    );
+}
+
+#[test]
+fn test_esp32dev_enriched_fields() {
+    let config = BoardConfig::from_board_id("esp32dev", &HashMap::new()).unwrap();
+    assert_eq!(config.core, "esp32");
+    assert_eq!(config.variant, "esp32");
+    // ESP32 boards now intentionally drop the JSON-shipped flash_mode
+    // (see comments in `from_board_id`). Downstream consumers fall back
+    // to mcu_config.default_flash_mode() which is "dio".
+    assert_eq!(config.flash_mode, None);
+    assert_eq!(config.memory_type, None);
+    assert_eq!(config.f_flash, Some("40000000L".to_string()));
+    assert_eq!(config.ldscript, Some("esp32_out.ld".to_string()));
+    assert_eq!(config.upload_speed, Some("460800".to_string()));
+}
+
+#[test]
+fn test_esp32_flash_mode_env_override_honoured() {
+    // The user can opt back into QIO via `board_build.flash_mode = qio`
+    // in their [env:X] section, which the daemon translates into a
+    // `flash_mode` override key.
+    let mut overrides = HashMap::new();
+    overrides.insert("flash_mode".to_string(), "qio".to_string());
+    let config = BoardConfig::from_board_id("esp32dev", &overrides).unwrap();
+    assert_eq!(config.flash_mode, Some("qio".to_string()));
+}
+
+#[test]
+fn test_esp32_flash_size_board_build_override_honoured() {
+    let mut overrides = HashMap::new();
+    overrides.insert("flash_size".to_string(), "4MB".to_string());
+
+    let config = BoardConfig::from_board_id("esp32-c6-devkitc-1", &overrides).unwrap();
+
+    assert_eq!(config.max_flash, Some(4 * 1024 * 1024));
+}
+
+#[test]
+fn test_esp32_flash_size_board_upload_override_honoured() {
+    let mut overrides = HashMap::new();
+    overrides.insert("upload.flash_size".to_string(), "4MB".to_string());
+
+    let config = BoardConfig::from_board_id("esp32-c6-devkitc-1", &overrides).unwrap();
+
+    assert_eq!(config.max_flash, Some(4 * 1024 * 1024));
+}
+
+#[test]
+fn test_boards_txt_upload_flash_size_sets_max_flash() {
+    let f = write_boards_txt(
+        "\
+demo.name=Demo ESP32-C6
+demo.build.mcu=esp32c6
+demo.build.f_cpu=160000000L
+demo.upload.flash_size=4MB
+",
+    );
+
+    let config = BoardConfig::from_boards_txt(f.path(), "demo", &HashMap::new()).unwrap();
+
+    assert_eq!(config.max_flash, Some(4 * 1024 * 1024));
+}
+
+#[test]
+fn test_nrf52840_dk_carries_arduino_macro() {
+    // Regression for FastLED CI fail: missing ARDUINO_NRF52_DK meant FastLED's
+    // nRF52 variants header fell through to the "Unknown variant" fallback
+    // (pin 1 invalid), blowing up Apa102.ino's static_assert. See fbuild#298.
+    let config = BoardConfig::from_board_id("nrf52840_dk", &HashMap::new()).unwrap();
+    let defines = config.get_defines();
+    assert_eq!(defines.get("ARDUINO_NRF52_DK"), Some(&"1".to_string()));
+}
+
+#[test]
+fn test_nrf52840_dk_adafruit_carries_arduino_macro() {
+    // Same fbuild#298 root cause — used by FastLED's `supermini_nrf52840` board.
+    let config = BoardConfig::from_board_id("nrf52840_dk_adafruit", &HashMap::new()).unwrap();
+    let defines = config.get_defines();
+    assert_eq!(
+        defines.get("ARDUINO_NRF52840_PCA10056"),
+        Some(&"1".to_string())
+    );
+    assert_eq!(defines.get("NRF52840_XXAA"), Some(&"1".to_string()));
+}
+
+#[test]
+fn test_adafruit_feather_nrf52840_sense_carries_arduino_macro() {
+    let config =
+        BoardConfig::from_board_id("adafruit_feather_nrf52840_sense", &HashMap::new()).unwrap();
+    let defines = config.get_defines();
+    assert_eq!(
+        defines.get("ARDUINO_NRF52840_FEATHER_SENSE"),
+        Some(&"1".to_string())
+    );
+    assert_eq!(defines.get("NRF52840_XXAA"), Some(&"1".to_string()));
+}
+
+/// Representative nRF52840 boards from the broader fbuild#298 sweep.
+/// PR #299 only added 3 priority boards; this verifies the bulk
+/// `enrich_boards` re-run covered the rest of the family.
+#[test]
+fn test_nrf52840_sweep_arduino_macros() {
+    // (board_id, expected macros that must appear as defines)
+    let cases: &[(&str, &[&str])] = &[
+        (
+            "adafruit_feather_nrf52840",
+            &["ARDUINO_NRF52840_FEATHER", "NRF52840_XXAA"],
+        ),
+        (
+            "adafruit_clue_nrf52840",
+            &["ARDUINO_NRF52840_CLUE", "NRF52840_XXAA"],
+        ),
+        (
+            "adafruit_itsybitsy_nrf52840",
+            &[
+                "ARDUINO_NRF52840_ITSYBITSY",
+                "ARDUINO_NRF52_ITSYBITSY",
+                "NRF52840_XXAA",
+            ],
+        ),
+        (
+            "adafruit_metro_nrf52840",
+            &[
+                "ARDUINO_NRF52840_FEATHER",
+                "ARDUINO_NRF52840_METRO",
+                "NRF52840_XXAA",
+            ],
+        ),
+        (
+            "adafruit_ledglasses_nrf52840",
+            &["ARDUINO_NRF52840_LED_GLASSES", "NRF52840_XXAA"],
+        ),
+        (
+            "adafruit_cplaynrf52840",
+            &["ARDUINO_NRF52840_CIRCUITPLAY", "NRF52840_XXAA"],
+        ),
+    ];
+    for (board_id, required) in cases {
+        let config = BoardConfig::from_board_id(board_id, &HashMap::new())
+            .unwrap_or_else(|e| panic!("failed to load board '{board_id}': {e}"));
+        let defines = config.get_defines();
+        for define in *required {
+            assert_eq!(
+                defines.get(*define),
+                Some(&"1".to_string()),
+                "board '{board_id}' missing -D{define} (got defines: {:?})",
+                defines.keys().collect::<Vec<_>>()
+            );
+        }
+    }
+}
+
+/// Boards from non-nRF52 platforms swept in fbuild#298. Each should now
+/// carry its Arduino board macro from upstream PlatformIO `build.extra_flags`.
+#[test]
+fn test_non_nrf52_sweep_arduino_macros() {
+    let cases: &[(&str, &[&str])] = &[
+        // atmelsam SAMD boards
+        ("zero", &["ARDUINO_SAMD_ZERO"]),
+        ("mkrzero", &["ARDUINO_SAMD_MKRZERO"]),
+        ("adafruit_feather_m0", &["ARDUINO_SAMD_FEATHER_M0"]),
+        // 8051 (intel_mcs51): board-id macro
+        ("AT89S51", &["AT89S51", "NAKED_ARCH_MCS51"]),
+        // PIC32 (microchippic32)
+        ("chipkit_uc32", &["_BOARD_UC32_"]),
+    ];
+    for (board_id, required) in cases {
+        let config = BoardConfig::from_board_id(board_id, &HashMap::new())
+            .unwrap_or_else(|e| panic!("failed to load board '{board_id}': {e}"));
+        let defines = config.get_defines();
+        for define in *required {
+            assert!(
+                defines.contains_key(*define),
+                "board '{board_id}' missing -D{define} (got defines: {:?})",
+                defines.keys().collect::<Vec<_>>()
+            );
+        }
+    }
+}
+
+#[test]
+fn test_pico_enriched_fields() {
+    let config = BoardConfig::from_board_id("rpipico", &HashMap::new()).unwrap();
+    assert_eq!(config.core, "earlephilhower");
+    assert_eq!(config.variant, "rpipico");
+    assert_eq!(config.upload_protocol, Some("picotool".to_string()));
+}
+
+#[test]
+fn test_sparkfun_xrp_controller_board_config() {
+    // SparkFun XRP Controller (RP2350B); maxgerhardt/platform-raspberrypi.
+    // Regression for FastLED `rp2350B SparkfunXRP` workflow (#295).
+    let config = BoardConfig::from_board_id("sparkfun_xrp_controller", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "rp2350");
+    assert_eq!(config.core, "earlephilhower");
+    assert_eq!(config.variant, "sparkfun_xrp_controller");
+    assert_eq!(config.upload_protocol, Some("picotool".to_string()));
+}
+
+#[test]
+fn test_extra_flags_produce_defines() {
+    // Enriched extra_flags should produce correct ARDUINO_* defines
+    let config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
+    let defines = config.get_defines();
+    assert_eq!(defines.get("ARDUINO_AVR_UNO"), Some(&"1".to_string()));
+}
+
+#[test]
+fn test_esp32c3_board_config() {
+    let config = BoardConfig::from_board_id("esp32c3", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "esp32c3");
+    assert_eq!(config.core, "esp32");
+    // ESP32 boards drop the JSON-shipped flash_mode (see
+    // `test_esp32dev_enriched_fields`); fall back is "dio" from MCU config.
+    assert_eq!(config.flash_mode, None);
+    assert_eq!(config.memory_type, None);
+    assert_eq!(config.ldscript, Some("esp32c3_out.ld".to_string()));
+    // ESP32-C3 DevKit runs at 160 MHz
+    assert_eq!(config.f_cpu, "160000000L");
+}
+
+#[test]
+fn test_esp32_effective_memory_type_tracks_effective_flash_mode() {
+    let config = BoardConfig::from_board_id("esp32c3", &HashMap::new()).unwrap();
+    assert_eq!(
+        config.effective_esp32_memory_type("dio"),
+        Some("dio_qspi".to_string())
+    );
+}
+
+#[test]
+fn test_esp32p4_evboard_uses_es_chip_variant() {
+    // The ESP32-P4 Function EV Board ships eco0–eco2 silicon ("ES pre rev.300").
+    // It must link against the `esp32p4_es` SDK (base ROM), not `esp32p4` (eco5
+    // ROM) — otherwise the bootloader panics on an illegal instruction.
+    let config = BoardConfig::from_board_id("esp32-p4-evboard", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "esp32p4");
+    assert_eq!(config.chip_variant, Some("esp32p4_es".to_string()));
+    assert_eq!(config.sdk_variant(), "esp32p4_es");
+}
+
+#[test]
+fn test_esp32p4_r3_uses_eco5_chip_variant() {
+    // The rev.300 board targets eco5 silicon and links the `esp32p4` SDK.
+    let config = BoardConfig::from_board_id("esp32-p4_r3", &HashMap::new()).unwrap();
+    assert_eq!(config.chip_variant, Some("esp32p4".to_string()));
+    assert_eq!(config.sdk_variant(), "esp32p4");
+}
+
+#[test]
+fn test_sdk_variant_falls_back_to_mcu() {
+    // Boards without an explicit chip_variant resolve the SDK dir from the MCU.
+    let config = BoardConfig::from_board_id("esp32c3", &HashMap::new()).unwrap();
+    assert_eq!(config.chip_variant, None);
+    assert_eq!(config.sdk_variant(), "esp32c3");
+}
+
+#[test]
+fn test_esp32_effective_memory_type_preserves_opi_flash_profiles() {
+    let config = BoardConfig::from_board_id("esp32-s3-devkitc-1-n32r8v", &HashMap::new()).unwrap();
+    assert_eq!(
+        config.effective_esp32_memory_type("dio"),
+        Some("opi_opi".to_string())
+    );
+}
+
+#[test]
+fn test_qemu_psram_config_absent_for_non_psram_board() {
+    let config = BoardConfig::from_board_id("esp32-s3-devkitc-1", &HashMap::new()).unwrap();
+    assert_eq!(config.qemu_esp32_psram_config(), None);
+}
+
+#[test]
+fn test_qemu_psram_config_detects_quad_psram_board() {
+    let config = BoardConfig::from_board_id("esp32-s3-devkitc1-n8r2", &HashMap::new()).unwrap();
+    assert_eq!(
+        config.qemu_esp32_psram_config(),
+        Some(Esp32QemuPsramConfig {
+            size_mib: 2,
+            is_octal: false,
+        })
+    );
+}
+
+#[test]
+fn test_qemu_psram_config_detects_octal_psram_board() {
+    let config = BoardConfig::from_board_id("esp32-s3-devkitc1-n8r8", &HashMap::new()).unwrap();
+    assert_eq!(
+        config.qemu_esp32_psram_config(),
+        Some(Esp32QemuPsramConfig {
+            size_mib: 8,
+            is_octal: true,
+        })
+    );
+}
+
+#[test]
+fn test_esp32c3_devkitm1_board_config() {
+    // Direct look up by full board ID (same underlying JSON)
+    let config = BoardConfig::from_board_id("esp32-c3-devkitm-1", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "esp32c3");
+    let flags = config.extra_flags.unwrap_or_default();
+    assert!(
+        flags.contains("ARDUINO_ESP32C3_DEV"),
+        "expected ARDUINO_ESP32C3_DEV in extra_flags, got: {flags}"
+    );
+}
+
+#[test]
+fn test_esp32c3_no_psram() {
+    // The plain C3 DevKit has no PSRAM
+    let config = BoardConfig::from_board_id("esp32c3", &HashMap::new()).unwrap();
+    let flags = config.extra_flags.clone().unwrap_or_default();
+    assert!(
+        !flags.contains("BOARD_HAS_PSRAM"),
+        "ESP32-C3 DevKit should not have PSRAM flag, got: {flags}"
+    );
+}
+
+#[test]
+fn test_attiny1604_board_config() {
+    let config = BoardConfig::from_board_id("ATtiny1604", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "attiny1604");
+    assert_eq!(config.core, "megatinycore");
+    assert_eq!(config.variant, "txy4");
+    assert_eq!(config.platform(), Some(fbuild_core::Platform::AtmelMegaAvr));
+    let defines = config.get_defines();
+    assert_eq!(defines.get("ARDUINO_ARCH_MEGAAVR"), Some(&"1".to_string()));
+}
+
+#[test]
+fn test_nano_every_board_config() {
+    let config = BoardConfig::from_board_id("nano_every", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "atmega4809");
+    assert_eq!(config.core, "arduino");
+    assert_eq!(config.variant, "nona4809");
+    assert_eq!(config.platform(), Some(fbuild_core::Platform::AtmelMegaAvr));
+}
+
+#[test]
+fn test_xiaoble_adafruit_board_config() {
+    // Seeed XIAO BLE (nRF52840, no IMU); maxgerhardt platform-nordicnrf52.
+    // Regression for FastLED `nrf52_xiaoblesense` workflow (#293).
+    let config = BoardConfig::from_board_id("xiaoble_adafruit", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "nrf52840");
+    assert_eq!(config.core, "nRF5");
+    assert_eq!(config.variant, "Seeed_XIAO_nRF52840");
+}
+
+#[test]
+fn test_xiaoblesense_adafruit_board_config() {
+    // Seeed XIAO BLE Sense (nRF52840 + IMU/mic).
+    // Regression for FastLED `adafruit_xiaoblesense` workflow (#293).
+    let config = BoardConfig::from_board_id("xiaoblesense_adafruit", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "nrf52840");
+    assert_eq!(config.core, "nRF5");
+    assert_eq!(config.variant, "Seeed_XIAO_nRF52840_Sense");
+}
+
+/// Validate that ALL megatinycore boards have the required framework-injected
+/// defines in extra_flags. PlatformIO's builder script injects these at build
+/// time, but fbuild must carry them in the board JSON since we don't run
+/// PlatformIO's SCons scripts.
+#[test]
+fn test_megatinycore_boards_have_required_defines() {
+    let db = get_board_db();
+    let required = &[
+        "MEGATINYCORE=",
+        "MEGATINYCORE_MAJOR=",
+        "MEGATINYCORE_MINOR=",
+        "MEGATINYCORE_PATCH=",
+        "MEGATINYCORE_RELEASED=",
+        "CORE_ATTACH_ALL",
+        "TWI_MORS",
+        "CLOCK_SOURCE=",
+    ];
+    let mut failures = Vec::new();
+    for (board_id, value) in db.iter() {
+        let core = value
+            .pointer("/build/core")
+            .and_then(|v| v.as_str())
+            .unwrap_or("");
+        if core != "megatinycore" {
+            continue;
+        }
+        let flags = value
+            .pointer("/build/extra_flags")
+            .and_then(|v| v.as_str())
+            .unwrap_or("");
+        for &define in required {
+            if !flags.contains(&format!("-D{define}")) {
+                failures.push(format!("{board_id}: missing -D{define}"));
+            }
+        }
+    }
+    assert!(
+        failures.is_empty(),
+        "megatinycore boards missing required framework defines:\n{}",
+        failures.join("\n")
+    );
+}
+
+/// Validate that ALL dxcore boards have the required framework-injected
+/// defines in extra_flags.
+#[test]
+fn test_dxcore_boards_have_required_defines() {
+    let db = get_board_db();
+    let required = &[
+        "DXCORE=",
+        "DXCORE_MAJOR=",
+        "DXCORE_MINOR=",
+        "DXCORE_PATCH=",
+        "DXCORE_RELEASED=",
+        "CORE_ATTACH_ALL",
+        "TWI_MORS_SINGLE",
+        "MILLIS_USE_TIMER",
+        "CLOCK_SOURCE=",
+    ];
+    let mut failures = Vec::new();
+    for (board_id, value) in db.iter() {
+        let core = value
+            .pointer("/build/core")
+            .and_then(|v| v.as_str())
+            .unwrap_or("");
+        if core != "dxcore" {
+            continue;
+        }
+        let flags = value
+            .pointer("/build/extra_flags")
+            .and_then(|v| v.as_str())
+            .unwrap_or("");
+        for &define in required {
+            if !flags.contains(&format!("-D{define}")) {
+                failures.push(format!("{board_id}: missing -D{define}"));
+            }
+        }
+    }
+    assert!(
+        failures.is_empty(),
+        "dxcore boards missing required framework defines:\n{}",
+        failures.join("\n")
+    );
+}
+
+#[test]
+fn test_uno_debug_tools_has_simavr() {
+    let config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
+    let tools = config
+        .debug_tools
+        .as_ref()
+        .expect("uno should have debug tools");
+    assert!(tools.contains_key("simavr"), "uno should have simavr");
+    assert!(tools.contains_key("avr-stub"), "uno should have avr-stub");
+    // simavr is not marked as onboard in the board JSON
+    assert!(!tools["simavr"].onboard);
+}
+
+#[test]
+fn test_emulators_filters_hardware_probes() {
+    let config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
+    let emus = config.emulators();
+    // simavr is an emulator, avr-stub is not in EMULATOR_TOOL_NAMES
+    assert!(emus.contains_key("simavr"), "simavr should be in emulators");
+    assert!(
+        !emus.contains_key("avr-stub"),
+        "avr-stub is not an emulator"
+    );
+}
+
+#[test]
+fn test_has_emulator() {
+    let config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
+    assert!(config.has_emulator("simavr"));
+    assert!(!config.has_emulator("qemu"));
+    assert!(!config.has_emulator("avr-stub")); // not in EMULATOR_TOOL_NAMES
+}
+
+#[test]
+fn test_debug_tools_round_trip_serde() {
+    let config = BoardConfig::from_board_id("uno", &HashMap::new()).unwrap();
+    let json = serde_json::to_string(&config).unwrap();
+    let restored: BoardConfig = serde_json::from_str(&json).unwrap();
+    assert_eq!(config.debug_tools, restored.debug_tools);
+}
+
+#[test]
+fn test_boards_txt_has_no_debug_tools() {
+    let f = write_boards_txt(
+        "\
+uno.name=Arduino Uno
+uno.build.mcu=atmega328p
+uno.build.f_cpu=16000000L
+",
+    );
+    let config = BoardConfig::from_boards_txt(f.path(), "uno", &HashMap::new()).unwrap();
+    assert!(
+        config.debug_tools.is_none(),
+        "boards.txt should not have debug tools"
+    );
+}
+
+#[test]
+fn test_esp32c6_devkitc1_has_variant() {
+    // Regression: esp32-c6-devkitc-1.json was missing build.variant,
+    // causing fbuild to look for variants/esp32-c6-devkitc-1/ instead
+    // of variants/esp32c6/, which broke compilation (pins_arduino.h
+    // not found). See https://github.com/FastLED/fbuild/issues/46
+    let config = BoardConfig::from_board_id("esp32-c6-devkitc-1", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "esp32c6");
+    assert_eq!(config.core, "esp32");
+    assert_eq!(
+        config.variant, "esp32c6",
+        "esp32-c6-devkitc-1 must have variant=esp32c6, not the board ID fallback"
+    );
+}
+
+#[test]
+fn test_esp32c6_alias_has_variant() {
+    // The 'esp32c6' alias resolves to esp32-c6-devkitm-1 which has
+    // the variant field. Verify both paths produce correct variant.
+    let config = BoardConfig::from_board_id("esp32c6", &HashMap::new()).unwrap();
+    assert_eq!(config.mcu, "esp32c6");
+    assert_eq!(
+        config.variant, "esp32c6",
+        "esp32c6 alias must resolve to variant=esp32c6"
+    );
+}
+
+#[test]
+fn test_board_without_debug_section() {
+    // Find a board that has no debug section (if any), or verify graceful handling
+    // by checking that debug_tools is populated from the JSON when present
+    let config = BoardConfig::from_board_id("esp32dev", &HashMap::new()).unwrap();
+    // esp32dev has debug tools (hardware probes only, no emulators)
+    if let Some(ref tools) = config.debug_tools {
+        let emus = config.emulators();
+        // esp32dev has no software emulators, only hardware probes
+        assert!(
+            emus.is_empty(),
+            "esp32dev should have no emulators, got: {:?}",
+            emus
+        );
+        // But it should still have hardware debug tools
+        assert!(!tools.is_empty(), "esp32dev should have debug tools");
+    }
+}
